lora_lite: use float throughout; build+flash the PortaPack firmware
The Cortex-M4F has only a single-precision FPU, so every double in the
demod was software-emulated (kilobytes of __aeabi_d*). Converting the
core to float shrinks the M4 baseband image enough for the external app
(UI + bundled baseband) to fit its 32 KB slot. Host decode is unchanged
(63 CRC-valid on capture.iq; 14/14 tests) since the FFT/cx path was
already float.

Also: -Os on proc_lorarx (baseband is -O3 globally), and a full
build-and-flash recipe (BUILD_AND_FLASH.md) that actually produces a
flashable 2 MB image with GCC 9.2.1. Regenerated the git-apply patch.

+++ lora_kiss.h
@@ -19,9 +19,9 @@ public:
         maxlog_ = max_log2;
         int maxN = 1 << max_log2;
         for (int i = 0; i < maxN; i++) {
-            double a = -2.0 * LORA_PI * i / maxN;   // full-size twiddles; subsampled per stage
-            tw_[i][0] = (float)cos(a);
-            tw_[i][1] = (float)sin(a);
+            float a = (float)(-2.0 * LORA_PI * i / maxN);  // twiddles; subsampled per stage
+            tw_[i][0] = cosf(a);
+            tw_[i][1] = sinf(a);
         }
         maxN_ = maxN;
     }
